+++ src/main/java/org/scion/jpan/internal/ScionHeaderParser.java
@@ -183,7 +183,11 @@ public class ScionHeaderParser {
 
     // rewind to original offset
     data.position(start);
-    return new InetSocketAddress(dstIP, dstPort);
+    try {
+      return new InetSocketAddress(dstIP, dstPort);
+    } catch (IllegalArgumentException e) {
+      throw e;
+    }
   }
 
   private static int extractDstPort(